refactor(run-query): added mock-data for mocking bigquery results
