--- test/queries/break-something.sql
@@ -1,4 +1,0 @@
---- helix-param: helix
---- helix-param2: helix2
---- helix-param3: helix3
-# this query is intentionally broken.
