--- test/queries/list-everything.sql
@@ -1,2 +1,0 @@
-select * 
-from requests201905
